Project-root/requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
psycopg[binary]==3.2.2

# Authentication & Security
Flask-Login==0.6.3
Flask-WTF==1.2.1
Authlib==1.3.0
requests==2.32.5
bcrypt==4.1.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Environment Management
python-dotenv==1.0.1


# Security Enhancements
Flask-Limiter==3.5.0
flask-talisman==1.1.0
python-magic-bin==0.4.14  # Required for secure file upload validation (includes libmagic for Windows)
redis==5.0.1

# CORS Support
Flask-CORS==4.0.0

# Performance
jsmin==3.0.1
cssmin==0.2.0

# Production WSGI Servers
gunicorn==21.2.0
waitress==3.0.0

# Testing
pytest==8.1.1
pytest-flask==1.3.0
pytest-cov>=4.0
ruff>=0.1
